+++ lib/constant.dart
@@ -38,8 +38,8 @@ extension TranslationTypeExtension on TranslationType {
 
 /// Audio Edition from Islamic Network
 enum AudioEdition {
-  /// ar.Abdulbasitmurattal
-  ar_Abdulbasitmurattal,
+  /// ar.abdulbasitmurattal
+  ar_abdulbasitmurattal,
 
   /// ar.abdullahbasfar
   ar_abdullahbasfar,
@@ -112,4 +112,23 @@ enum AudioEdition {
 
   /// ar.aymanswoaid
   ar_aymanswoaid;
+}
+
+/// An error message
+class ErrorMessages {
+  /// Error for Juz number out of range
+  static String juzNumberOutOfRange = "Juz number out of range";
+
+  /// Error for chapter number out of range
+  static String chapterNumberOutOfRange = "Chapter number out of range";
+
+  /// Error for verse number out of range
+  static String verseNumberOutOfRange = "Verse number out of range";
+
+  /// Error for chapter number and verse number invalid
+  static String chapterNumberAndVerseNumberInvalid =
+      "Invalid chapter number and/or verse number";
+
+  /// Error for verse key invalid
+  static String verseKeyInvalid = "Invalid verse key";
 }